structure_understand/requirements.txt
PyYAML>=6.0
requests>=2.30
openai>=1.0
PyPDF2>=3.0
pytest>=8.0
ollama